Track model indices using Item#getIndex
Previous method of tracking fails when an Item is added to the
ExpandableView and another item (any but the last one) removed.  The
issue manifests because Item objects are reused, together with the
IndexTrackingModel objects kept inside of Item objects. However, a list
of IndexTrackingModels kept inside of ExpandableView gets desynchronized
with IndexTrackingModels kept inside of Items.